refactor-#35: refactor es module
