+++ src/elastic-search/config/index.ts
@@ -1,0 +1,1 @@
+export * from './elastic-search-config.service';
