+++ src/elastic-search/helpers/convertRecordBodyToEsString.ts
