+++ src/elastic-search/interfaces/search-body.interface.ts
