+++ src/elastic-search/interfaces/search-result.interface.ts
