+++ src/elastic-search/services/index.ts
@@ -1,0 +1,1 @@
+export * from './elastic-search.service';
