--- src/es/config/index.ts
@@ -1,1 +1,0 @@
-export * from './es-async-options';
